requirements.txt
PyMuPDF
paddlepaddle
<<<<<<< HEAD
paddleocr
=======
paddleocr
Pillow
<<<<<<< HEAD
python-Levenshtein
=======
>>>>>>> origin/main
>>>>>>> bd703b3 (20 integrate dbnet and implement cer measurement (#21))
